Deps: Replace node-sass with sass
Node sass has been deprecated in 2020 in favor of a new dart-sass compiler implementation. We replace the dependency with the dart implementation.

Doing this removes the dependency on `node-gyp` and thus python2.

We needed to fix a single deprecation error from the dart compiler (https://sass-lang.com/documentation/at-rules/extend#disallowed-selectors).

This fix did not have any visible effect (subtitles are still grey) which I confirmed in a firefox build.